refactor: general cleanup
