+++ www/.eslintrc.json
@@ -1,0 +1,8 @@
+{
+  "globals": {
+    "graphql": true
+  },
+  "rules": {
+    "react/prop-types": [0]
+  }
+}
