+++ www/src/views/showcase/collabsible.js
@@ -1,0 +1,153 @@
+import React, { Component } from "react"
+import hex2rgba from "hex2rgba"
+import { style } from "glamor"
+import FaAngleDown from "react-icons/lib/fa/angle-down"
+import FaAngleUp from "react-icons/lib/fa/angle-up"
+
+import { options, scale, rhythm } from "../../utils/typography"
+import presets, { colors } from "../../utils/presets"
+
+class Collapsible extends Component {
+  state = {
+    collapsed: false,
+  }
+
+  handleClick = () => {
+    this.setState({ collapsed: !this.state.collapsed })
+  }
+
+  render() {
+    const { heading, children } = this.props
+    return (
+      <div>
+        {/* TODO: onClick should be on a link or something */}
+        <h4
+          css={{
+            color: colors.lilac,
+            cursor: `pointer`,
+            display: `flex`,
+            alignItems: `center`,
+            fontWeight: `normal`,
+            fontSize: scale(-2 / 5).fontSize,
+            marginTop: rhythm(options.blockMarginBottom),
+            marginRight: rhythm(5 / 4),
+            letterSpacing: `.15em`,
+            textTransform: `uppercase`,
+            "&:hover": {
+              color: colors.gatsby,
+            },
+          }}
+          onClick={this.handleClick}
+        >
+          {heading}
+          {` `}
+          <span css={{ marginLeft: `auto` }}>
+            {this.state.collapsed ? <FaAngleDown /> : <FaAngleUp />}
+          </span>
+        </h4>
+        <div
+          css={{
+            overflow: `hidden`,
+            height: this.state.collapsed ? `0px` : `500px`,
+            transition: `height 0.2s`,
+          }}
+        >
+          <div
+            css={{
+              overflowY: `scroll`,
+              height: `500px`,
+              ...styles.scrollbar,
+            }}
+          >
+            {children}
+          </div>
+        </div>
+      </div>
+    )
+  }
+}
+
+export default Collapsible
+
+const styles = {
+  featuredSitesCard: style({
+    display: `flex`,
+    flexDirection: `column`,
+    flexGrow: 0,
+    flexShrink: 0,
+    width: 320,
+    marginBottom: rhythm(options.blockMarginBottom * 2),
+    marginRight: rhythm(3 / 4),
+    [presets.Hd]: {
+      width: 360,
+      marginRight: rhythm(6 / 4),
+    },
+    [presets.VHd]: {
+      width: 400,
+    },
+  }),
+  withTitleHover: style({
+    "& .title": {
+      transition: `box-shadow .3s cubic-bezier(.4,0,.2,1), transform .3s cubic-bezier(.4,0,.2,1)`,
+      boxShadow: `inset 0 0px 0px 0px ${colors.ui.whisper}`,
+    },
+    "&:hover .title": {
+      boxShadow: `inset 0 -3px 0px 0px ${colors.ui.bright}`,
+    },
+  }),
+  button: {
+    border: 0,
+    borderRadius: presets.radius,
+    cursor: `pointer`,
+    fontFamily: options.headerFontFamily.join(`,`),
+    fontWeight: `bold`,
+    padding: `${rhythm(1 / 5)} ${rhythm(2 / 3)}`,
+    WebkitFontSmoothing: `antialiased`,
+    "&&": {
+      backgroundColor: colors.gatsby,
+      borderBottom: `none`,
+      boxShadow: `none`,
+      color: `white`,
+      "&:hover": {
+        backgroundColor: colors.gatsby,
+      },
+    },
+  },
+  sticky: {
+    paddingTop: rhythm(options.blockMarginBottom),
+    position: `sticky`,
+    top: 0,
+    [presets.Desktop]: {
+      top: `calc(${presets.headerHeight} - 1px)`,
+    },
+  },
+  scrollbar: {
+    WebkitOverflowScrolling: `touch`,
+    "&::-webkit-scrollbar": {
+      width: `6px`,
+      height: `6px`,
+    },
+    "&::-webkit-scrollbar-thumb": {
+      background: colors.ui.bright,
+    },
+    "&::-webkit-scrollbar-track": {
+      background: colors.ui.light,
+    },
+  },
+  screenshot: {
+    borderRadius: presets.radius,
+    boxShadow: `0 4px 10px ${hex2rgba(colors.gatsby, 0.1)}`,
+    marginBottom: rhythm(options.blockMarginBottom / 2),
+    transition: `all ${presets.animation.speedDefault} ${
+      presets.animation.curveDefault
+    }`,
+  },
+  screenshotHover: {
+    background: `transparent`,
+    color: colors.gatsby,
+    "& .gatsby-image-wrapper": {
+      transform: `translateY(-3px)`,
+      boxShadow: `0 8px 20px ${hex2rgba(colors.lilac, 0.5)}`,
+    },
+  },
+}
